% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_best_learner.R
\name{default_learners}
\alias{default_learners}
\title{Creates a list of learners}
\usage{
default_learners()
}
\value{
a named \code{list} of learners, where each element consists of a
\itemize{
\item \code{model}: A \code{parsnip} model specification
\item \code{grid}: A \code{data.frame} with columns corresponding to tuning parameters
}
}
\description{
This function creates a list of learners compatible with the \code{learners}
argument of \link{fit_best_learner}, which is used as the default argument.
}
\examples{
default_learners()

}
