% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rctglm_methods.R
\name{estimand}
\alias{estimand}
\title{Generic for extracting estimand}
\usage{
estimand(x, ...)
}
\arguments{
\item{x}{an object of some class to dispatch on}

\item{...}{additional arguments passed to methods}
}
\description{
Generic for extracting estimand
}
