% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superlearner.R
\name{fit_best_learner}
\alias{fit_best_learner}
\title{Find the best learner in terms of RMSE among specified learners using cross validation}
\usage{
fit_best_learner(formula, data, n_folds = 5, learners = default_learners())
}
\arguments{
\item{formula}{an object of class \code{"\link[stats]{formula}"} (or one that
    can be coerced to that class): a symbolic description of the
    model to be fitted.  The details of model specification are given
    under \sQuote{Details}.}

\item{data}{an optional data frame, list or environment (or object
    coercible by \code{\link{as.data.frame}} to a data frame) containing
    the variables in the model.  If not found in \code{data}, the
    variables are taken from \code{environment(formula)},
    typically the environment from which \code{glm} is called.}

\item{n_folds}{a \code{numeric} with the number of cross-validation folds used when fitting and
evaluating models}

\item{learners}{a \code{list} of \code{tidymodels}. Default uses a combination of MARS, linear
regression and boosted trees}
}
\value{
a trained \code{workflow}
}
\description{
Find the best learner in terms of RMSE among specified learners using cross validation
}
\examples{
# Generate some "historical" data of a control treatment
n <- 100
w1 <- runif(n, min = -2, max = 2)
x1 <- abs(sin(w1))
a <- rbinom (n, 1, .5)
b0 <- 1
b1 <- 1.5
b2 <- 2

truemean_notreat <- b0+b1*x1
y_notreat <- rnorm(n, mean = truemean_notreat, sd = 1)
dat_hist <- data.frame(Y = y_notreat, W = w1)

# Fit a learner to the data with default learners
fit <- fit_best_learner(Y ~ ., data = dat_hist)
}
