% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate pairs of historical and current data}
\usage{
generate_data(
  ATE = 3,
  ATE.shift = rnorm(1, 0, 0.1),
  N.covs = 10,
  N.overspec = 0,
  coefs = c(0.5, 1, 1),
  mu.shift = 0,
  cov = diag(1, N.covs + N.overspec),
  noise = 1,
  N.hist.control = 5000,
  N.hist.treatment = 100,
  N.test.control = 0,
  N.test.treatment = 0,
  N.control = 200,
  N.treatment = 300,
  N.sim = 1000,
  N.cores = 16
)
}
\arguments{
\item{ATE}{The average treatment effect \eqn{ATE = \mathbb{E}[Y(1)] - \mathbb{E}[Y(0)] } data is generated from}

\item{ATE.shift}{ATE shift in the historical data, when historical data contains treatment group patients}

\item{N.covs}{Number of covariates in the data generating process, indexed as x1,...,x_N.covs}

\item{N.overspec}{Number of overspecified covariates (generates covariates not used in the data generating process, but can be used in model specifications to simulate overspecification). They are indexed as x_{N.covs+1},..., x_{N.covs+N.overspec}.}

\item{coefs}{Coefficients for a, b and c in \eqn{Y(w) \sim N(aX^TX+bX+cXW+ATE\cdot W)}}

\item{mu.shift}{Mean shift in the historical data "away" from mean 0 in rct data (to simulate skewed historical distribution relative to RCT data).}

\item{cov}{Covariance structure of normally distributed covariates (simultaneous between "true" and overspecified covariates)}

\item{noise}{Std. err of error term in generated data}

\item{N.hist.control}{Number of patients in historical control groups}

\item{N.hist.treatment}{Number of patients in historical treatment groups}

\item{N.test.control}{Additional number of historical control group patients having same distribution as "hist" (if provided, these can be used for preliminary power estimation; see power.est argument for lm.hist)}

\item{N.test.treatment}{Additional number of historical treatment group patients having same distribution as "hist" (if provided, these can be used for preliminary power estimation; see power.est argument for lm.hist)}

\item{N.control}{Number of patients in current RCT control groups}

\item{N.treatment}{Number of patients in current RCT treatment groups}

\item{N.sim}{Number of generated historical and RCT datasets}

\item{N.cores}{Number of cores to use for parallelisation}
}
\value{
List object consisting of $hist (historical data), $hist_test (test data
with same distribution of the historical data), $rct (current RCT data)
and multiple attributes.
}
\description{
Current and historical data simulated from a normal distribution as \eqn{Y(w) \sim N(aX^TX+bX+cXW+ATE\cdot W)}
}
\examples{
generate_data(N.sim = 1, N.hist.control = 10, N.hist.treatment = 0,
              N.control = 1, N.treatment = 1)

}
