% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.hist.R
\name{lasso.hist}
\alias{lasso.hist}
\title{LASSO regression as function of historical data}
\usage{
lasso.hist(data.hist, method.covs = c("x1", "x2"))
}
\arguments{
\item{data.hist}{Data.frame of the historical data used to build the regression model.}

\item{method.covs}{Character vector with names of the covariates to use as baseline covariates in the prediction model. Make sure that categorical variables are considered as factors.}
}
\value{
An object of class "cv.glmnet" is returned, which is a list with
the ingredients of the cross-validation fit. See further details of
glmnet::cv.glmnet.
}
\description{
See glmn::cv.glmnet for further details
}
\examples{
data <- sim.lm(N.sim = 1, N.hist.control = 100, N.hist.treatment = 100,
              N.control = 50, N.treatment = 50)

lasso.hist(data[[1]]$hist)


}
