% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.hist.R
\name{lm.hist}
\alias{lm.hist}
\title{ANCOVA models using historical data}
\usage{
lm.hist(
  data.list,
  method = c("None", "PROCOVA", "PSM"),
  sup.margin = 0,
  alpha = 0.05,
  outcome.var = "y",
  treatment.var = "w",
  adj.covs = NULL,
  pred.model = NULL,
  interaction = FALSE
)
}
\arguments{
\item{data.list}{A list of elements $hist and $rct, which are both data.frames being historical and current RCT data sets, respectively.}

\item{method}{Method for using historical data. Options: None, PROCOVA, PSM. None refers to ANOVA (for adj.covs = NULL) and ANCOVA (for adj.covs specified).}

\item{sup.margin}{Superiority margin (for non-inferiority margin, a negative value can be provided).}

\item{alpha}{Significance level. Due to regulatory guidelines when using a one-sided test, half the specified significance level is used. Thus, for standard alpha = .05, a significance level of 2.5\\% is used.}

\item{outcome.var}{Character with the name of the outcome variable in both the $rct and $hist data set.}

\item{treatment.var}{Character with the name of the outcome treatment indicator in both the $rct and $hist data set. Notice that the treatment variable should be an indicator with treatment == 1 and control == 0.}

\item{adj.covs}{Character vector with names of the covariates to adjust for as raw covariates in the ANCOVA model for estimating the ATE. Make sure that categorical variables are considered as factors.}

\item{pred.model}{Model object which should be a function of the historical data that fits a prediction model based on the baseline covariates. This is only needed for method == "PROCOVA" or method == "PSM". The model object obtained from the function should be a valid argument for \link[stats]{predict}.}

\item{interaction}{Logical value, that determines whether to model interaction effects between covariates and treatment indicator when estimating the ATE. For method = "PROCOVA", the prognostic score is regarded as an additional covariate and thus the interaction between the prognostic score and the treatment indicator is included.}
}
\value{
lm.hist returns an object of \link[base]{class} "lm". The functions summary and \code{\link[stats:anova]{stats::anova()}} are used to obtain and print
a summary and analysis of variance table of the results. The generic accessor functions coefficients, effects, fitted.values and residuals
extract various useful features of the value returned by lm. The object contains a list of the same components as an object of class "lm"
but with the extra component $test_sup.margin, which is a list of the  t-test statistic value 'test_stat' and the result of the t-test 'test_result'
based on the superiority margin. Look at \href{https://www.rdocumentation.org/packages/stats/versions/3.6.2/topics/lm}{lm} for further information on the class.
}
\description{
The function fits ANCOVA models using the function \link[stats]{lm} but potentially leveraging a historical data set.
}
