% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.psm.R
\name{lm.psm}
\alias{lm.psm}
\title{PSM for randomised clinical trials}
\usage{
lm.psm(
  data.list,
  margin = 0,
  alpha = 0.05,
  outcome.var = "y",
  treatment.var = "w",
  adj.covs = NULL,
  interaction = FALSE,
  B = 100,
  ...
)
}
\description{
The function estimates the ATE using a procedure for one-to-one propensity score matching in the
setting of randomised clinical studies where data from multiple external control groups are available.
The procedure involves matching the patients in the treatment group to historical controls on
observed confounders, as well as estimating the outcome bias between control groups induced
by unobserved confounders. It is then possible to use this estimated bias between control groups
with the aim of removing the bias in the estimated treatment effect by implicitly taking these
unobserved confounders into account. The procedure was described by Lim et al. (Lim J, Walley R, Yuan J, Liu J, Dabral A, Best N, et al. Minimizing Patient Burden
Through the Use of Historical Subject-Level Data in Innovative Confirmatory Clinical Trials:
Review of Methods and Opportunities. Therapeutic Innovation & Regulatory Science. 2018;52(5):546–559.),
ensuring that all control arm patients from the current RCT are used in the treatment effect estimation in equation
\deqn{\widehat{ATT} = \frac{1}{n_1} \sum_{i:w_i=1} \left( y_i(1) - \widehat y_{i,j}(0) \right)}
}
\details{
The describtion of the procedure can be found in further details in the article by Lim et al.
}
