% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.wider.R
\name{param.wider}
\alias{param.wider}
\title{Transform data to wide format and extracting baseline parameters}
\usage{
param.wider(
  data,
  visit_name,
  visit_n,
  param_name,
  parameters,
  analysis_value,
  subject_id,
  analysis_day
)
}
\arguments{
\item{data}{The data set that should be transformed to wide format}

\item{visit_name}{Name of the column with the visit numbers}

\item{visit_n}{Visit number for the randomisation visit}

\item{param_name}{Name of the column with the parameter names}

\item{parameters}{Vector of parameters that should be pivoted to their own column}

\item{analysis_value}{Name of column with the value of the parameter}

\item{subject_id}{Name of the column with subject identification}

\item{analysis_day}{Name of column with the relative analysis date}
}
\value{
Data set in wide format with a column for each entrance of the parameters vector.
}
\description{
Function that take a data set and write it to wide format and take baseline measurement as the latest
measurement at or before baseline.
}
\examples{
data <- data.frame(subject_id = rep(1:3, rep(4,3)), visit = rep(1:2,
rep(2,2)), parameter = rep(c("measure1", "measure2"), 6), analysis_val =
rnorm(12, 100, 3))[1:10, ]

param.wider(data = data, visit_name = visit, visit_n = 2, param_name =
parameter, parameters = c("measure1", "measure2"), analysis_value =
analysis_val, subject_id = subject_id, analysis_day = NULL)

}
