% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.ancova.R
\name{power.ancova}
\alias{power.ancova}
\title{Estimation of prospective power for ANCOVA models}
\usage{
power.ancova(
  data.hist,
  outcome.var = "y",
  treatment.var = "w",
  adj.covs = NULL,
  interaction = NULL,
  n,
  r,
  ATE,
  margin = 0,
  alpha = 0.05
)
}
\arguments{
\item{data.hist}{Data.frame of the historical data used to estimate sigma, rho, and R2 in the prospective power estimation.}

\item{outcome.var}{Character with the name of the outcome variable data.hist.}

\item{treatment.var}{Character with the name of the treatment variable data.hist.}

\item{adj.covs}{Character vector with names of the covariates to adjust for as raw covariates in the ANCOVA model for estimating the ATE. Make sure that categorical variables are considered as factors.}

\item{interaction}{Logical value, that determines whether to model interaction effects between covariates and treatment indicator when estimating R2.}

\item{n}{Number of participants in total in the new data set. From this number of participants in the treatment group is \eqn{n1=(r/(1+r))n} and the control group is \eqn{n1=(1/(1+r))n}.}

\item{r}{Allocation ratio \eqn{r=n1/n0}. For one-to-one randomisation r=1.}

\item{ATE}{Minimum effect size that we should be able to detect.}

\item{margin}{Superiority margin (for non-inferiority margin, a negative value can be provided).}
}
\value{
Vector of the estimated entities sigma, rho/R2, power_NC, and power_GS.
}
\description{
Estimation of prospective power for ANCOVA models
}
\details{
This function calculates the prospective power based on theoretical non-centrality parameter as well as Guenther-Schouten approximations,
using historical data provided in data.hist. The entities sigma, rho, and R2 are calculated using the historical data, and these can not
be specified by the user. Look at power.NC or power.GS if you want a power calculation based on user specified entities.
In addition, the necessary entities for these calculations are outputted.

When wanting to estimate the power for the PROCOVA method add a column to data.hist with the predicted values using the prognostic model and add the
name of this column to the adj.covs. Note that data.hist should be independent of the data used to build the the prognostic model to obtain accurate
power estimations.
}
