% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.random.R
\name{predict.random}
\alias{predict.random}
\title{Predict for class random}
\usage{
\method{predict}{random}(ran, newdata)
}
\arguments{
\item{ran}{object of class random.}

\item{newdata}{The new data that}
}
\value{
Returns a numeric object with the "predicted" values.
}
\description{
Predict for class random
}
