% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rctglm_prog_methods.R
\name{prog_model}
\alias{prog_model}
\alias{prog_model.rctglm_prog}
\title{Generic for extracting the fitted prognostic model}
\usage{
prog_model(x)

\method{prog_model}{rctglm_prog}(x)
}
\arguments{
\item{x}{an object of class \code{rctglm_prog} (returned by
\link{rctglm_with_prognosticscore})}
}
\description{
Extract the fitted prognostic model.
}
\examples{
# Generate some data
n <- 100
b0 <- 1
b1 <- 1.5
b2 <- 2
W1 <- runif(n, min = -2, max = 2)

dat_treat <- glm_data(
  b0+b1*abs(sin(W1))+b2*A,
  W1 = W1,
  A = rbinom (n, 1, .5)
)

dat_notreat <- glm_data(
  b0+b1*abs(sin(W1)),
  W1 = W1
)

ate <- rctglm_with_prognosticscore(
  formula = Y ~ .,
  group_indicator = A,
  data = dat_treat,
  family = gaussian(),
  estimand_fun = "ate",
  data_hist = dat_notreat)

prog_model(ate)
}
