% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.hist.R
\name{random.hist}
\alias{random.hist}
\title{Find range of outcome variable
This function can be used to generate a random prediction model, in the sense that it
determines the range of the outcome variable in the historical data, and returns this
range as a object of class "random". Then predict.random randomly predicts the outcome
within this range.}
\usage{
random.hist(data.hist)
}
\arguments{
\item{data.hist}{Data.frame of the historical data used to find the range of the outcome variable in order to randomly predict a value.}
}
\value{
Returns a vector of the minimum and maximum value in data.hist with the class "random".
}
\description{
Find range of outcome variable
This function can be used to generate a random prediction model, in the sense that it
determines the range of the outcome variable in the historical data, and returns this
range as a object of class "random". Then predict.random randomly predicts the outcome
within this range.
}
