% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rctglm_methods.R
\name{se_estimand}
\alias{se_estimand}
\alias{se_estimand.rctglm}
\title{Generic for extracting standard error (SE) of estimand}
\usage{
se_estimand(x, ...)

\method{se_estimand}{rctglm}(x)
}
\arguments{
\item{x}{an object of some class to dispatch on}

\item{...}{additional arguments passed to methods}
}
\description{
Extract the standard error (SE) of the estimand. See more details in
\link{rctglm} and \link{estimand}.
This function is a shortcut to extract the list element \code{se_estimand} of
an \code{rctglm} class object.
}
\examples{
# Generate some data to showcase example
n <- 100
dat_binom <- glm_data(
  1+1.5*X1+2*A,
  X1 = rnorm(n),
  A = rbinom(n, 1, .5),
  family = binomial()
)

# Fit the model
ate <- rctglm(formula = Y ~ .,
              group_indicator = A,
              data = dat_binom,
              family = binomial,
              estimand_fun = "ate")

se_estimand(ate)
}
