% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{variance_bound_gs}
\alias{variance_bound_gs}
\title{Title}
\usage{
variance_bound_gs(data, response = Y, ..., inflation = 1.25, deflation = 1)
}
\arguments{
\item{deflation}{}
}
\description{
Title
}
\details{
See article Zimmermann, G., Kieser, M., & Bathke, A. C. (2019). Sample size
calculation and blinded recalculation for analysis of covariance models with
multiple random covariates. Journal of Biopharmaceutical Statistics, 30(1),
143–159. https://doi.org/10.1080/10543406.2019.1632871
}
