% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_options.R
\name{options}
\alias{options}
\title{PostCard Options}
\arguments{
\item{verbose}{Verbosity level specifying how much information should be printed to the user. Allowed values are 1 and 0 (Defaults to \code{1}, overwritable using option 'PostCard.verbose' or environment variable 'R_POSTCARD_VERBOSE')}
}
\description{
Internally used, package-specific options. All options will prioritize R options() values, and fall back to environment variables if undefined. If neither the option nor the environment variable is set, a default value is used.
}
\section{Checking Option Values}{

Option values specific to \code{PostCard} can be
accessed by passing the package name to \code{env}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{options::opts(env = "PostCard")

options::opt(x, default, env = "PostCard")
}\if{html}{\out{</div>}}
}

\section{Options}{

\describe{
\item{verbose}{\describe{
Verbosity level specifying how much information should be printed to the user. Allowed values are 1 and 0\item{default: }{\preformatted{1}}
\item{option: }{PostCard.verbose}
\item{envvar: }{R_POSTCARD_VERBOSE (evaluated if possible, raw string otherwise)}
}}

}
}

\seealso{
options getOption Sys.setenv Sys.getenv
}
