% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_plot.R
\name{plot_power_marginaleffect}
\alias{plot_power_marginaleffect}
\title{Plot power curves calculated using \code{power_marginaleffect} for a list of models}
\usage{
plot_power_marginaleffect(
  target_effect,
  exposure_prob,
  model_list = default_power_model_list(),
  test_data_fun = function(n) {
     glm_data(Y ~ 1 + 3 * sin(W)^2 + 2 * X, W = runif(n,
    min = -2, max = 2), X = rnorm(n))
 },
  desired_power = 0.9,
  ns = 10:250,
  n_iter = 1,
  ...
)
}
\arguments{
\item{target_effect}{Passed to \code{\link[=power_marginaleffect]{power_marginaleffect()}}}

\item{exposure_prob}{Passed to \code{\link[=power_marginaleffect]{power_marginaleffect()}}}

\item{model_list}{a named \code{list} of models used to get predictions on generated test
data sets that are then passed to \code{\link[=power_marginaleffect]{power_marginaleffect()}} as \code{predictions}}

\item{test_data_fun}{a \code{function} with a single argument \code{n} that generates test
data sets for the sample sizes \code{ns} specified}

\item{desired_power}{a \code{numeric} between 0 and 1 indicating the desired power level}

\item{ns}{a \code{numeric} vector of sample sizes}

\item{n_iter}{a \code{numeric} indicating a number of iterations to process and average over}

\item{...}{additional arguments passed to \code{\link[=power_marginaleffect]{power_marginaleffect()}}}
}
\value{
a \code{ggplot2} object
}
\description{
Plot power curves calculated using \code{power_marginaleffect} for a list of models
}
\examples{
# A simple use case with default models and test data
plot_power_marginaleffect(target_effect = 1.3, exposure_prob = 0.5)

# Specify a margin with the ellipsis argument
plot_power_marginaleffect(target_effect = 1.3, exposure_prob = 0.5, margin = 1.3)
}
