% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcard-package.R
\docType{package}
\name{postcard-package}
\alias{postcard}
\alias{postcard-package}
\title{postcard: Estimating Marginal Effects with Prognostic Covariate Adjustment}
\description{
Conduct power analyses and inference of marginal effects. Uses plug-in estimation and influence functions to perform robust inference, optionally leveraging historical data to increase precision with prognostic covariate adjustment. The methods are described in Højbjerre-Frandsen et al. (2025) \doi{10.48550/arXiv.2503.22284}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://novonordisk-opensource.github.io/postcard/}
  \item \url{https://github.com/NovoNordisk-OpenSource/postcard}
  \item Report bugs at \url{https://github.com/NovoNordisk-OpenSource/postcard/issues}
}

}
\author{
\strong{Maintainer}: Mathias Lerbech Jeppesen \email{mathiasljeppesen@outlook.com}

Authors:
\itemize{
  \item Emilie Hojbjerre-Frandsen \email{ehfd@novonordisk.com}
}

Other contributors:
\itemize{
  \item Novo Nordisk A/S [copyright holder]
}

}
\keyword{internal}
