% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostCard-package.R
\docType{package}
\name{PostCard-package}
\alias{PostCard}
\alias{PostCard-package}
\title{PostCard: Prognostic covariate adjustment for estimation of average treatment effect}
\description{
Convenient functions for conducting analysis using prognostic covariate adjustment for linear models. The package contain a function that can generate normally distributed data in different data generating scenarios in order to determine the finite sample properties of prognostic covariate for linear models. This generated data can then be used as an input for the lm.hist.sim function, that estimates ATE, SD, power, type I error, etc. This function can also be used for a single pair of RCT and historical data.
}
\author{
\strong{Maintainer}: Emilie Hoejbjerre-Frandsen \email{ehfd@novonordisk.com}

}
\keyword{internal}
