% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostCard-package.R
\docType{package}
\name{PostCard-package}
\alias{PostCard}
\alias{PostCard-package}
\title{PostCard: Estimation of Marginal Effects Using Prognostic Covariate Adjustment}
\description{
Conducting analysis using robust plug-in GLM estimation with influence functions and prognostic covariate adjustment.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nnpackages.github.io/PostCard/}
}

}
\author{
\strong{Maintainer}: Mathias Lerbech Jeppesen \email{mathiasljeppesen@outlook.com}

Authors:
\itemize{
  \item Emilie Hojbjerre-Frandsen \email{ehfd@novonordisk.com}
}

}
\keyword{internal}
