% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcard-package.R
\docType{package}
\name{postcard-package}
\alias{postcard}
\alias{postcard-package}
\title{postcard: Estimation of Marginal Effects Using Prognostic Covariate Adjustment}
\description{
Conducting analysis using robust plug-in GLM estimation with influence functions and prognostic covariate adjustment.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nnpackages.github.io/postcard/}
}

}
\author{
\strong{Maintainer}: Mathias Lerbech Jeppesen \email{mathiasljeppesen@outlook.com}

Authors:
\itemize{
  \item Emilie Hojbjerre-Frandsen \email{ehfd@novonordisk.com} [copyright holder]
}

}
\keyword{internal}
