% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostCard-package.R
\docType{package}
\name{PostCard-package}
\alias{PostCard}
\alias{PostCard-package}
\title{PostCard: Prognostic covariate adjustment for estimation of estimands}
\description{
Convenient functions for conducting analysis using prognostic covariate adjustment for GLMs. The package uses plug-in estimation for robust estimation of the estimand and influence functions for robust estimation of the variance of the estimand.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nnpackages.github.io/PostCard/}
}

}
\author{
\strong{Maintainer}: Mathias Lerbech Jeppesen \email{mathiasljeppesen@outlook.com}

Authors:
\itemize{
  \item Emilie Hojbjerre-Frandsen \email{ehfd@novonordisk.com}
}

}
\keyword{internal}
