% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rctglm_prog_methods.R
\name{prog}
\alias{prog}
\alias{prog.rctglm_prog}
\title{Extract information about the fitted prognostic model}
\usage{
prog(x)

\method{prog}{rctglm_prog}(x)
}
\arguments{
\item{x}{an object of class \code{rctglm_prog} (returned by
\link{rctglm_with_prognosticscore})}
}
\value{
description
}
\description{
Extracts the \code{prognostic_info} list element from an \code{rctglm_prog} object. See
'Value' at \link{rctglm_with_prognosticscore} for more details.
}
\examples{
# Generate some data
n <- 100
b0 <- 1
b1 <- 1.5
b2 <- 2
W1 <- runif(n, min = -2, max = 2)

dat_treat <- glm_data(
  Y ~ b0+b1*abs(sin(W1))+b2*A,
  W1 = W1,
  A = rbinom (n, 1, .5)
)

dat_notreat <- glm_data(
  Y ~ b0+b1*abs(sin(W1)),
  W1 = W1
)

ate <- rctglm_with_prognosticscore(
  formula = Y ~ .,
  group_indicator = A,
  data = dat_treat,
  family = gaussian(),
  estimand_fun = "ate",
  data_hist = dat_notreat)

prog(ate)
}
