% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rctglm.R
\name{rctglm}
\alias{rctglm}
\title{Fit GLM and find any estimand (marginal effect) using plug-in estimation with variance estimation using
influence functions}
\usage{
rctglm(
  formula,
  family,
  data,
  group_indicator,
  group_allocation_prob = 1/2,
  estimand_fun = "ate",
  estimand_fun_deriv0 = NULL,
  estimand_fun_deriv1 = NULL,
  ...
)
}
\arguments{
\item{formula}{an object of class \code{"\link[stats]{formula}"} (or one that
    can be coerced to that class): a symbolic description of the
    model to be fitted.  The details of model specification are given
    under \sQuote{Details}.}

\item{family}{a description of the error distribution and link
    function to be used in the model.  For \code{glm} this can be a
    character string naming a family function, a family function or the
    result of a call to a family function.  For \code{glm.fit} only the
    third option is supported.  (See \code{\link[stats]{family}} for details of
    family functions.)}

\item{data}{an optional data frame, list or environment (or object
    coercible by \code{\link{as.data.frame}} to a data frame) containing
    the variables in the model.  If not found in \code{data}, the
    variables are taken from \code{environment(formula)},
    typically the environment from which \code{glm} is called.}

\item{group_indicator}{(name of) the variable in data that identifies randomisation groups}

\item{group_allocation_prob}{a \code{numeric} with the probabiliy of being assigned "group 1" (rather than group 0)}

\item{estimand_fun}{a \code{character} specifying a default estimand function, or \code{function} with
arguments \code{psi0} and \code{psi1} specifying the estimand.}

\item{estimand_fun_deriv0}{a \code{function} specifying the derivative of \code{estimand_fun} wrt. \code{psi0}}

\item{estimand_fun_deriv1}{a \code{function} specifying the derivative of \code{estimand_fun} wrt. \code{psi1}}

\item{...}{Additional arguments passed to \code{\link[stats:glm]{stats::glm()}}}
}
\value{
an \code{rctglm} object
}
\description{
Fit GLM and find any estimand (marginal effect) using plug-in estimation with variance estimation using
influence functions
}
\details{
Default estimand functions can be specified via \code{"ate"} and \code{"rate_ratio"}.

As a default, the \code{Deriv} package to perform symbolic differentiation to find the derivatives of
the \code{estimand_fun}.
}
\examples{
# Generate some data
n <- 100
x1 <- rnorm (n)
a <- rbinom (n, 1, .5)
b0 <- 1
b1 <- 1.5
b2 <- 2
lin_pred <- b0+b1*x1+b2*a

y_norm <- rnorm(n, mean = lin_pred, sd = 1)
dat_norm <- data.frame(Y = y_norm, X = x1, A = a)

ate <- rctglm(formula = Y ~ .,
              group_indicator = A,
              data = dat_norm,
              family = gaussian(),
              estimand_fun = "ate")
}
