% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rctglm.R
\name{rctglm}
\alias{rctglm}
\title{Fit GLM and find any estimand (marginal effect) using plug-in estimation with variance estimation using
influence functions}
\usage{
rctglm(
  formula,
  group_indicator,
  family,
  data,
  group_allocation_prob = NULL,
  estimand_fun = "ate",
  estimand_fun_deriv0 = NULL,
  estimand_fun_deriv1 = NULL,
  cv_variance = TRUE,
  cv_variance_folds = 5,
  verbose = options::opt("verbose"),
  ...
)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class):
a symbolic description of the model to be fitted. The details of model specification are
given under ‘Details’ in the \link{glm} documentation.}

\item{group_indicator}{(name of) the \emph{binary} variable in \code{data} that
identifies randomisation groups. The variable is required to be binary to
make the "orientation" of the \code{estimand_fun} clear.}

\item{family}{a description of the error distribution and link
    function to be used in the model.  For \code{glm} this can be a
    character string naming a family function, a family function or the
    result of a call to a family function.  For \code{glm.fit} only the
    third option is supported.  (See \code{\link[stats]{family}} for details of
    family functions.)}

\item{data}{an optional data frame, list or environment (or object
    coercible by \code{\link{as.data.frame}} to a data frame) containing
    the variables in the model.  If not found in \code{data}, the
    variables are taken from \code{environment(formula)},
    typically the environment from which \code{glm} is called.}

\item{group_allocation_prob}{a \code{numeric} with the probabiliy of being assigned "group 1" (rather than group 0).
As a default, the ratio of 1's in data is used.}

\item{estimand_fun}{a \code{function} with arguments \code{psi0} and \code{psi1} specifying
the estimand. Alternative, specify "ate" or "rate_ratio" as a \code{character}
to use one of the default estimand functions. See
more details in the "Estimand" section of this documentation.}

\item{estimand_fun_deriv0}{a \code{function} specifying the derivative of \code{estimand_fun} wrt. \code{psi0}. As a default
the algorithm will use symbolic differentiation to automatically find the derivative from \code{estimand_fun}}

\item{estimand_fun_deriv1}{a \code{function} specifying the derivative of \code{estimand_fun} wrt. \code{psi1}. As a default
the algorithm will use symbolic differentiation to automatically find the derivative from \code{estimand_fun}}

\item{cv_variance}{a \code{logical} determining whether to estimate the variance
using cross-validation (see details of \link{rctglm}).}

\item{cv_variance_folds}{a \code{numeric} with the number of folds to use for cross
validation if \code{cv_variance} is \code{TRUE}.}

\item{verbose}{\code{numeric} verbosity level. Higher values means more information is
printed in console. A value of 0 means nothing is printed to console during
execution (Defaults to \code{2}, overwritable using option 'PostCard.verbose' or environment variable 'R_POSTCARD_VERBOSE')}

\item{...}{Additional arguments passed to \code{\link[stats:glm]{stats::glm()}}}
}
\value{
\code{rctglm} returns an object of class inheriting from \code{"rctglm"}.

The function \link{estimand} (or short-hand version \link{est}) can be used to extract
a \code{data.frame} with an estimated value and standard error of the estimand.

A method for the generic \link{coef} has been added for \code{rctglm}
(i.e., \link{coef.rctglm}), which uses the method \code{coef.glm} to extract coefficient
information from the underlying \code{glm} fit in the procedure.

An object of class \code{rctglm} is a list containing the following components:
\itemize{
\item \code{estimand}: A \code{data.frame} with plug-in estimate of estimand, standard
error (SE) estimate and variance estimate of estimand
\item \code{estimand_fun}: The \code{function} used to obtain an estimate of the estimand
from counterfactual means
\item \code{means_counterfactual}: A \code{data.frame} with counterfactual means \code{psi0} and \code{psi1}
\item \code{fitted.values_counterfactual}: A \code{data.frame} with counterfactual mean
values, obtained by transforming the linear predictors for each group
by the inverse of the link function.
\item \code{glm}: A \code{glm} object returned from running \link[stats:glm]{stats::glm} within the procedure
\item \code{call}: The matched \code{call}
}
}
\description{
The procedure uses plug-in-estimation and influence functions to perform robust inference of any specified
estimand in the setting of a randomised clinical trial, even in the case of heterogeneous effect of
covariates in randomisation groups.
}
\details{
The procedure assumes the setup of a randomised clinical trial with observations grouped by a binary
\code{group_indicator} variable, allocated randomly with probability \code{group_allocation_prob}. A GLM is
fit and then used to predict the response of all observations in the event that the \code{group_indicator}
is 0 and 1, respectively. Taking means of these predictions produce the \emph{counterfactual means}
\code{psi0} and \code{psi1}, and an estimand \code{r(psi0, psi1)} is calculated using any specified \code{estimand_fun}.

The variance of the estimand is found by taking the variance of the influence function of the estimand.
If \code{cv_variance} is \code{TRUE}, then the counterfactual predictions for each observation (which are
used to calculate the value of the influence function) is obtained as out-of-sample (OOS) predictions
using cross validation with number of folds specified by \code{cv_variance_folds}. The cross validation splits
are performed using stratified sampling with \code{group_indicator} as the \code{strata} argument in \link[rsample:vfold_cv]{rsample::vfold_cv}.

This method of inference using plug-in estimation and influence functions for the variance produces a
causal estimate of the estimand, as stated by articles XXXX.
}
\section{Estimands}{

As noted in the description, \code{psi0} and \code{psi1} are the counterfactual means found by prediction using
a fitted GLM in the binary groups defined by \code{group_indicator}.

Default estimand functions can be specified via \code{"ate"} (which uses the function
\code{function(psi1, psi0) psi1-psi0}) and \code{"rate_ratio"} (which uses the function
\code{function(psi1, psi0) psi1/psi0}). See more information on specifying the \code{estimand_fun}
in section
\href{https://nnpackages.github.io/PostCard/articles/more-details.html#specifying-any-estimand}{Specifying any estimand}
in the vignette \code{vignette("more-details")}.

As a default, the \code{Deriv} package is used to perform symbolic differentiation to find the derivatives of
the \code{estimand_fun}.
}

\examples{
# Generate some data to showcase example
n <- 100
dat_gaus <- glm_data(
  Y ~ 1+1.5*X1+2*A,
  X1 = rnorm(n),
  A = rbinom(n, 1, .5),
  family = gaussian()
)

# Fit the model
ate <- rctglm(formula = Y ~ .,
              group_indicator = A,
              data = dat_gaus,
              family = gaussian)

# Pull information on estimand
estimand(ate)

## Another example with different family and specification of estimand_fun
dat_binom <- glm_data(
  Y ~ 1+1.5*X1+2*A,
  X1 = rnorm(n),
  A = rbinom(n, 1, .5),
  family = binomial()
)

rr <- rctglm(formula = Y ~ .,
              group_indicator = A,
              data = dat_binom,
              family = binomial(),
              estimand_fun = "rate_ratio")

odds_ratio <- function(psi1, psi0) (psi1*(1-psi0))/(psi0*(1-psi1))
or <- rctglm(formula = Y ~ .,
              group_indicator = A,
              data = dat_binom,
              family = binomial,
              estimand_fun = odds_ratio)


}
