% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rctglm_methods.R
\name{rctglm_methods}
\alias{rctglm_methods}
\alias{estimand}
\alias{estimand.rctglm}
\alias{est}
\alias{coef.rctglm}
\alias{print.rctglm}
\title{Methods for objects of class \code{rctglm}}
\usage{
estimand(object)

\method{estimand}{rctglm}(object)

est(object)

\method{coef}{rctglm}(object, ...)

\method{print}{rctglm}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{object}{an object of class \code{rctglm}}

\item{...}{additional arguments passed to methods}

\item{x}{an object of class \code{rctglm}}

\item{digits}{a \code{numeric} with the number of digits to display when printing}
}
\description{
Methods mostly to extract information from model fit and inference. See
details for more information on each method.
}
\details{
\code{estimand} and short-hand form \code{est} are methods for extracting the estimated
estimand value as well as the standard error (SE) and variance of the estimand.
These functions are a shortcut to extract the list element \code{estimand}
of an \code{rctglm} class object.

\code{coef} just use the corresponding \code{glm} methods on the \code{glm}
fit contained within the \code{rctglm} object. Thus, this function is a
shortcuts to running \code{coef(x$glm)}.
}
\examples{
# Generate some data to showcase example
n <- 100
dat_binom <- glm_data(
  1+1.5*X1+2*A,
  X1 = rnorm(n),
  A = rbinom(n, 1, .5),
  family = binomial()
)

# Fit the model
ate <- rctglm(formula = Y ~ .,
              exposure_indicator = A,
              data = dat_binom,
              family = binomial,
              estimand_fun = "ate")

print(ate)
estimand(ate)
coef(ate)
}
