% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rctglm_methods.R
\name{rctglm_methods}
\alias{rctglm_methods}
\alias{print.rctglm}
\alias{coef.rctglm}
\alias{summary.rctglm}
\alias{estimand}
\alias{estimand.rctglm}
\alias{se_estimand}
\alias{se_estimand.rctglm}
\title{Methods for objects of class \code{rctglm}}
\usage{
\method{print}{rctglm}(x, digits = 3, ...)

\method{coef}{rctglm}(object, ...)

\method{summary}{rctglm}(object, ...)

estimand(x)

\method{estimand}{rctglm}(x)

se_estimand(x)

\method{se_estimand}{rctglm}(x)
}
\arguments{
\item{x}{an object of class \code{rctglm}}

\item{digits}{a \code{numeric} with the number of digits to display when printing}

\item{...}{additional arguments passed to methods}

\item{object}{an object of class \code{rctglm}}
}
\description{
Methods mostly to extract information from model fit and inference. See
details for more information on each method.
}
\details{
\code{coef} and \code{summary} just use the corresponding \code{glm} methods on the \code{glm}
fit contained within the \code{rctglm} object. Thus, these functions are shortcuts
of running \code{coef(x$glm)} and \code{summary(x$glm)}.

\code{estimand} and \code{se_estimand} are methods for extracting the estimated
estimand value as well as the standard error (SE) of the estimand.
Extract the plug-in estimation of the estimand, found by using the
estimand function on the predicted counterfactual means of each group.
These functions are a shortcut to extract the list elements \code{estimand}
and \code{se_estimand}of an \code{rctglm} class object.
}
\examples{
# Generate some data to showcase example
n <- 100
dat_binom <- glm_data(
  1+1.5*X1+2*A,
  X1 = rnorm(n),
  A = rbinom(n, 1, .5),
  family = binomial()
)

# Fit the model
ate <- rctglm(formula = Y ~ .,
              group_indicator = A,
              data = dat_binom,
              family = binomial,
              estimand_fun = "ate")

print(ate)
estimand(ate)
se_estimand(ate)
coef(ate)
summary(ate)
}
