% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rctglm_with_prognosticscore.R
\name{rctglm_with_prognosticscore}
\alias{rctglm_with_prognosticscore}
\title{Title}
\usage{
rctglm_with_prognosticscore(
  formula,
  family,
  data,
  group_indicator,
  group_allocation_prob = 1/2,
  estimand_fun = "ate",
  estimand_fun_deriv0 = NULL,
  estimand_fun_deriv1 = NULL,
  ...,
  data_hist,
  prog_formula = NULL,
  n_folds = 5,
  learners = default_learners()
)
}
\arguments{
\item{formula}{an object of class \code{"\link[stats]{formula}"} (or one that
    can be coerced to that class): a symbolic description of the
    model to be fitted.  The details of model specification are given
    under \sQuote{Details}.}

\item{family}{a description of the error distribution and link
    function to be used in the model.  For \code{glm} this can be a
    character string naming a family function, a family function or the
    result of a call to a family function.  For \code{glm.fit} only the
    third option is supported.  (See \code{\link[stats]{family}} for details of
    family functions.)}

\item{data}{an optional data frame, list or environment (or object
    coercible by \code{\link{as.data.frame}} to a data frame) containing
    the variables in the model.  If not found in \code{data}, the
    variables are taken from \code{environment(formula)},
    typically the environment from which \code{glm} is called.}

\item{group_indicator}{(name of) the variable in data that identifies randomisation groups}

\item{group_allocation_prob}{a \code{numeric} with the probabiliy of being assigned "group 1" (rather than group 0)}

\item{estimand_fun}{a \code{character} specifying a default estimand function, or \code{function} with
arguments \code{psi0} and \code{psi1} specifying the estimand.}

\item{estimand_fun_deriv0}{a \code{function} specifying the derivative of \code{estimand_fun} wrt. \code{psi0}}

\item{estimand_fun_deriv1}{a \code{function} specifying the derivative of \code{estimand_fun} wrt. \code{psi1}}

\item{...}{Additional arguments passed to \code{\link[stats:glm]{stats::glm()}}}

\item{data_hist}{a \code{data.frame} with historical data on which to fit a prognostic model}

\item{prog_formula}{a \code{character} or \code{numeric} with the formula for fitting the prognostic
model on the historical data \code{data_hist}. Default models the response (assumed same as in
\code{formula}) using all columns in the \code{data_hist} data}

\item{n_folds}{a \code{numeric} with the number of cross-validation folds used when fitting and
evaluating models}

\item{learners}{a \code{list} of \code{tidymodels}. Default uses a combination of MARS, linear
regression and boosted trees}
}
\value{
an \code{rctglm} object, fitted with a prognostic score as a covariate in the model
}
\description{
Title
}
\details{
More details on prognostic models and scores being predictions of counterfactual means
in control group.
}
\examples{
# Generate some data
n <- 100
w1 <- runif(n, min = -2, max = 2)
x1 <- abs(sin(w1))
a <- rbinom (n, 1, .5)
b0 <- 1
b1 <- 1.5
b2 <- 2

truemean_treat <- b0+b1*x1+b2*a
y_treat <- rnorm(n, mean = truemean_treat, sd = 1)
dat_treat <- data.frame(Y = y_norm, W = w1, A = a)

truemean_notreat <- b0+b1*x1
y_notreat <- rnorm(n, mean = truemean_notreat, sd = 1)
dat_hist <- data.frame(Y = y_notreat, W = w1)

ate <- rctglm_with_prognosticscore(
  formula = Y ~ .,
  group_indicator = A,
  data = dat_norm,
  family = gaussian(),
  estimand_fun = "ate",
  data_hist = dat_hist)

}
