% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_plot.R
\name{repeat_power_linear}
\alias{repeat_power_linear}
\title{Create data of power curves calculated using functions in \code{\link[=power_linear]{power_linear()}} for a list of formulas/models}
\usage{
repeat_power_linear(
  ate,
  formula_list,
  train_data,
  power_fun = c("power_gs", "power_nc"),
  ns = 10:250,
  desired_power = 0.9,
  ...
)
}
\arguments{
\item{ate}{Passed to \code{\link[=power_gs]{power_gs()}} or \code{\link[=power_nc]{power_nc()}}}

\item{formula_list}{a named \code{list} of formulas that are element wise passed to
\code{\link[=variance_ancova]{variance_ancova()}}}

\item{train_data}{Passed as the \code{data} argument in \code{\link[=variance_ancova]{variance_ancova()}}}

\item{power_fun}{a \code{character} string with value \code{"power_gs"} or \code{"power_nc"},
specifying what function in the \code{\link[=power_linear]{power_linear()}} topic to use}

\item{ns}{a \code{numeric} vector of sample sizes}

\item{desired_power}{a \code{numeric} between 0 and 1 indicating the desired power level}

\item{...}{Arguments passed to \code{\link[=variance_ancova]{variance_ancova()}} and \code{\link[=power_gs]{power_gs()}} or \code{\link[=power_nc]{power_nc()}}}
}
\value{
a \code{ggplot2} object
}
\description{
Estimate a variance for power approximation using \code{\link[=variance_ancova]{variance_ancova()}} for each formula
in \code{formula_list} on \code{train_data}. Then calculate power using the function with name
specified in \code{power_fun} across a number of sample sizes \code{ns} for an assumed average
treatment effect of \code{ate}.
}
\examples{
train_data <- glm_data(
  Y ~ 1+3*sin(W)^2,
  W = runif(1e3, min = -2, max = 2)
)
repeat_power_linear(
  ate = 1.3, formula_list = list(ANCOVA = Y ~ W, ANCOVA2 = Y ~ 1),
  train_data = train_data)
}
