% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeat_power_plot.R
\name{repeat_power_marginaleffect}
\alias{repeat_power_marginaleffect}
\alias{plot.postcard_rpm}
\title{Create data and plot power curves calculated using \code{\link[=power_marginaleffect]{power_marginaleffect()}} for a list of models}
\usage{
repeat_power_marginaleffect(
  target_effect,
  exposure_prob,
  model_list = default_power_model_list(),
  test_data_fun = function(n) {
     glm_data(Y ~ 1 + 3 * sin(W)^2, W = stats::runif(n,
    min = -2, max = 2))
 },
  ns = seq(5, 100, 5),
  desired_power = 0.9,
  n_iter = 1,
  ...
)

\method{plot}{postcard_rpm}(x, cols = NULL, ...)
}
\arguments{
\item{target_effect}{Passed to \code{\link[=power_marginaleffect]{power_marginaleffect()}}}

\item{exposure_prob}{Passed to \code{\link[=power_marginaleffect]{power_marginaleffect()}}}

\item{model_list}{a named \code{list} of models used to get predictions on generated test
data sets that are then passed to \code{\link[=power_marginaleffect]{power_marginaleffect()}} as \code{predictions}. The
elements of \code{model_list} need to have an existing \code{predict()} method. The default is
an ANCOVA and a prognostic model fitted with \code{\link[=fit_best_learner]{fit_best_learner()}} to a simple data set
of 1000 observations generated with a non-linear effect of a single covariate using
\code{\link[=glm_data]{glm_data()}}.}

\item{test_data_fun}{a \code{function} with a single argument \code{n} that generates test
data sets for the sample sizes \code{ns} specified. The default generates data using
\code{\link[=glm_data]{glm_data()}} with the same data generating process as the training
data used to fit the default models in \code{model_list}.}

\item{ns}{a \code{numeric} vector of sample sizes}

\item{desired_power}{a \code{numeric} between 0 and 1 indicating the desired power level}

\item{n_iter}{a \code{numeric} indicating a number of iterations to process and average over}

\item{...}{additional arguments passed to \code{\link[=power_marginaleffect]{power_marginaleffect()}}}

\item{x}{an object of class \code{postcard_rpm} created by \code{repeat_power_marginaleffect()}}

\item{cols}{a (potentially named) \code{character} vector of colors for the different models
in \code{model_list}}
}
\value{
\code{repeat_power_marginal} returns an object of class \code{postcard_rpm}, which is
just a \code{data.frame} with a \code{plot} method defined. The \code{plot} method returns a
\code{ggplot2} object.
}
\description{
Iterate a process of simulating test data from \code{test_data_fun}, making predictions
using models in \code{model_list}, and calculating power using \code{\link[=power_marginaleffect]{power_marginaleffect()}}
across a number of sample sizes \code{ns} and iterations \code{n_iter}. The results are averaged
and used to create a plot of the resulting power curves.
}
\examples{
# A simple use case with default models and test data
rpm1305 <- repeat_power_marginaleffect(target_effect = 1.3, exposure_prob = 0.5)

# Specify a margin with the ellipsis argument
repeat_power_marginaleffect(target_effect = 1.3, exposure_prob = 0.5, margin = 1.3)

}
\seealso{
\code{\link[=repeat_power_linear]{repeat_power_linear()}} for a similar implementation to iterate the process
of approximating power with the functions in \code{\link[=power_linear]{power_linear()}}
}
