% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{variance_bound_gs}
\alias{variance_bound_gs}
\title{Title}
\usage{
variance_bound_gs(formula, data, inflation = 1.25, deflation = 1)
}
\arguments{
\item{data}{the data used to estimate the variance bound}

\item{deflation}{}

\item{response}{the response variable in the data for which}
}
\description{
Title
}
\details{
See article Zimmermann, G., Kieser, M., & Bathke, A. C. (2019). Sample size
calculation and blinded recalculation for analysis of covariance models with
multiple random covariates. Journal of Biopharmaceutical Statistics, 30(1),
143–159. https://doi.org/10.1080/10543406.2019.1632871
}
\examples{
# Generate a negative binomial response
nb <- glm_data(1+2*x1-x2,
                x1 = rnorm(10),
                x2 = rgamma(10, shape = 2),
                family = MASS::negative.binomial(2))
variance_bound_gs(nb, response_name = "Y")
}
